checkout git at specified tag
The script was modified by ChatGPT in response to the following prompt:
-
Add a version tag for each as well.
-

I'm surprised it understood my intent so well from the little that I said.

In addition to these changes, it helpfully provided me with a script to get
the latest tags, but I didn't need that, because my assumption is that I'll
update the tags only in response to any CVE's that are reported (or as part
of a regular update process).

Generating that API caller script skewed ChatGPT's world view though, because
it then misunderstood my next request which was:
-
Change the tags to the latest stable versions.
-

which caused it to rewrite the code with the addition of the GitHub API to
fetch the latest tags so I skipped past the changes it generated in response
to that.