Use defconfig instead of menuconfig
The script was modified by ChatGPT in response to the following prompt:
-
Use defconfig instead of menuconfig
-

I'm again blown away by the fact that it was able to understand my intent
even from the terse commands that I'm using. That set my expectations too
high I guess because it didn't quite grasp the next set of commands I gave
it, probably because it assumed I was still in exploratory mode.

To the prompt of:
-
Can busybox be extended to add git, make and gcc as well?
Effectively everything you would get with ubuntu build-essentials.
-

it responded just as it would to someone asking a newbie question by
babbling on about configure, make, make install (on alpine of all things,
then followed that up by talking about running
	apt install build-essential git
on ubuntu), and when I followed that up with this prompt:
-
Add this to the earlier perl script with git based pull
-

it again mistook that to mean adding calls to apt (it even showed the
output from running apt install). So I had to clarify what I meant and
that goes into the next commit.