build tools such as git, make and gcc as part of the script
As mentioned in the previous commit, ChatGPT misunderstood some of the
earlier prompts I gave it so I had to clarify and make my intent clearer
this time. As before, the script was modified by ChatGPT in response to
the following prompt:
-
No, I want the git repos of git, gcc and make to also be pulled and built
by the perl script (in addition to linux and busybox)
-

which seemed to be very clear instructions that it could then follow.