additional entries to match Linux From Scratch (LFS)
The script was modified by ChatGPT in response to the following prompt:
-
Add more entries so that all of the stuff built by linux from scratch
is part of the build.
-

It seems to have picked up a decent subset (but not all) of them,

At this point, I assume ChatGPT was hitting some internally imposed token
limit because it elided the tail end of the script so I just added the
changed part of the code manually.