more correct git links and tags
Rather than depend only on ChatGPT, I decided to see if there are other
AI providers who have equivalent functionality (without having to login)

Searching on duckduckgo gave a list of ~5 providers and of those Perplexity.AI
seemed to generate pretty good code, so I decided to check it out.

The first prompt I gave it was:
-
List the git links corresponding to each of the packages built as part of Linux From Scratch (LFS)
-

It responded with an overview rather than providing actual git links so I
prompted it again saying:
-
Try harder
-

which it took to heart because it then gave a fairly comprehensive list of
packages but checking against the wget list from LFS showed it was missing
the very first one: acl, so I prompted it again saying:
-
What about the missing stuff such as acl?
-

After looking through the list it generated, I then gave the following prompt:
-
List each of these in the following format with the tag as well:

"coreutils" => {
	url  => "https://github.com/coreutils/coreutils.git",
	tag  => "v8.32",  \# Specify the version you want
},
-

The list it generated seemed to more closely match the stuff in LFS 12.3
so I decided to merge the code from ChatGPT and Perplexity.AI

I wrote the merge script myself, but in hindsight I should have vibe coded
that as well but given that Perplexity.AI generated code that looked like this:
   "gawk": {
        url: "git://git.savannah.gnu.org/gawk.git",
        tag: "v5.3.0"
    }
ie JSON, not Perl code, I decided to just do it myself. The merged structure
has entries where duplicated keys use an array containing both entries since
I'm unsure which is correct although given that Perplexity.AI shows newer
versions of the tags (and the more correct git links), I plan to shift the
rest of my vibe coding over to Perplexity.AI