Generate a table of packages - manually, this time.
So far, during this exercise, I've not bothered to check if the stuff
generated by ChatGPT/PerplexityAI matches reality. Since I was "vibe
coding" I let it add whatever it wanted. This week I decided to sit
down and actually check if the stuff it generated matched reality
and as expected, things look pretty bad.

To start off, LFS 12.3 has a list of 94 elements in its wget-list.
The list generated by all of my prompting to ChatGPT/PerpelexityAI
with suggestions to "add more stuff" and to check its work resulted
in just 57 entries. So that is a fairly large group of 37 missing
packages - so 57/94 gives ~60% "truthiness".

For the packages that it did list, many had the wrong versions. Some
of the packages for which it got the correct versions were: groff,
mpc, sed, zlib, patch, tar, gmp, m4, mpfr, grep, make, attr, flex,
bzip2 and autoconf - 15 out of 57 for a "truthiness" rating of
(57/94)*(15/57) == ~16%

In addition to the sins of omission, it added some made up packages
that don't exist or it was confused about the names, for eg: awk, in
addition to gawk and procps/procps-ng, net-tools, etc.

The worst part of all was the list of links to the git repos that it
generated. The entire point of this "vibe coding" exercise was not to
make the coding easier - that is the most trivial part in some sense.

The real point of getting ChatGPT/PerplexityAI involved was to see if
the grunt work of mapping from the packages to their git repos could
be automated or at least made easier. And this is where it completely
fell short. Most of the git repos that were generated do not exist
so the data it generated is either old or atrophied due to link rot.

So I don't have much choice except to give AI an 'F' for this exercise.

I'm fairly disappointed but also glad that it appears to be making some
progress. Since I had to cross check the work anyway, I've generated a
table (of space separated values - hence the ssv tag to the file name)
of packages, version, download link and a git repo link / TODO marker
of all of the stuff in the wget-list from LFS 12.3 and this is the
canonical reference I'll use going forward.

The tail end of this list has 6 patches that probably won't have any
git repos associated with them, but I'm leaving them in here to match
the wget-list from LFS 12.3

All the packages with a git repo tagged with git:// are the GNU software
repos that I've downloaded (hence, verified already) and the rest are
marked TODO for now.